refactor: PostCard 반응형 수정
